Added Justify features and effects.
white-effect | Adds a white transition effect. (0.3s)
red-effect | Adds a red transition effect. (0.3s)
orange-effect | Adds a orange transition effect. (0.3s)
yellow-effect | Adds a yellow transition effect. (0.3s)

jcenter | This style attribute is used to center the elements in the flex container horizontally (on the x-axis).
jend | This style property aligns items inside the flex container horizontally (on the x-axis) to the right.
jright | Aligns the contents of the flex container horizontally to the right.
jleft | Aligns the contents of the flex container containing the class horizontally to the left.
jbetween | This style property aligns elements in the flex container horizontally (on the x-axis) by spacing them apart.